pyproject.toml
[project]
name = "tele-be"
version = "0.1.0"
description = "Atama poka poka, aaho bakaaa..."
readme = "README.md"
requires-python = ">=3.13.5"
dependencies = [
    "django>=5.2.7",
    "psycopg[binary]>=3.2.9",
    "django-cors-headers>=4.7.0",
    "django-filter==24.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.1",
    "whitenoise>=6.9.0",
]

[dependency-groups]
dev = [
    "django-silk>=5.4.2",
    "django-stubs>=5.2.2",
    "djangorestframework-stubs>=3.16.2",
    "ruff>=0.12.10",
]
extra = [
    "pynvim>=0.5.2",
]

[tool.mypy]
plugins = [
	"mypy_django_plugin.main",
	"mypy_drf_plugin.main"
]

[tool.ruff]
exclude = [
	".git",
	".venv",
	"migrations",
	"install"
]
line-length=79

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.format]
indent-style="space"
